backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.1

# Data handling
pydantic==2.9.2
PyYAML==6.0.2

# Media processing
Pillow==10.4.0

# Utility / Automation
watchdog==5.0.2
requests==2.32.3

# Optional: for database indexing (if you switch from JSON to SQLite later)
sqlalchemy==2.0.34
